requirements.txt
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
pymongo==4.6.1
python-dotenv==1.0.0
pandas==2.1.4
neo4j==5.15.0